Run differential rhythmicity analysis for RNA-seq data using DESeq2

Run differential rhythmicity analysis for RNA-seq data using DESeq2

Usage

compareRhythms_deseq2(
  counts,
  exp_design,
  lengths,
  period,
  rhythm_fdr,
  compare_fdr,
  amp_cutoff,
  just_classify
)

Arguments

exp_design

A data.frame of the experimental design with at least two columns: "time" and "group".

lengths

A data.frame of average transcript lengths. Only used with methods "deseq" and "edgeR".

period

The period of rhythm being tested (default = 24)

rhythm_fdr

The false discovery cutoff for finding rhythmic time series (default = 0.05)

compare_fdr

The false discovery cutoff for the comparison of rhythms (default = 0.05)

amp_cutoff

The minimum peak-to-trough amplitude in log2 scale considered biologically relevant (default = 0.5)

just_classify

Boolean specifying whether genes must only be classified (TRUE) or if the amplitude and phases of fits should also be returned (FALSE)